+++ lua/neg/groups/plugins/heirline.lua
@@ -1,0 +1,7 @@
+-- heirline.nvim — provide neutral groups for users to link
+return {
+  HeirlineNormal = { link = 'StatusLine' },
+  HeirlineInactive = { link = 'StatusLineNC' },
+  HeirlineWinbar = { link = 'WinBar' },
+}
+