+++ lua/neg/groups/plugins/ufo.lua
@@ -1,0 +1,7 @@
+-- nvim-ufo — folds
+return {
+  UfoFoldedBg = { link = 'Folded' },
+  UfoFoldedFg = { link = 'Folded' },
+  UfoCursorFoldedLine = { link = 'CursorLine' },
+}
+